% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mapof}
\alias{mapof}
\title{Download maps}
\usage{
mapof(x = c("state", "municip"))
}
\arguments{
\item{x}{x = 'state' or x = 'municip'.}
}
\value{
State border or the municipal bounderis.
}
\description{
Download maps
}
