% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municip}
\alias{municip}
\title{The map of Republic of Kosovo - Municipality borders}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 38 rows and 7 columns.
}
\source{
Original file https://github.com/justinelliotmeyers/Kosovo_2018_Administrative_Boundaries
}
\usage{
municip
}
\description{
The map of Republic of Kosovo - Municipality borders
}
\keyword{datasets}
