% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state}
\alias{state}
\title{The map of Republic of Kosovo - State border}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 8 columns.
}
\source{
OSM through QGIS.
}
\usage{
state
}
\description{
The map of Republic of Kosovo - State border
}
\keyword{datasets}
